backend/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
amqp==2.6.1               # via kombu
asgiref==3.2.10           # via django
billiard==3.6.3.0         # via celery
celery==4.4.1             # via -r requirements.in, django-celery-beat
certifi==2020.12.5        # via requests
chardet==4.0.0            # via requests
coreapi==2.3.3            # via drf-yasg
coreschema==0.0.4         # via coreapi, drf-yasg
django-celery-beat==2.2.0  # via -r requirements.in
django-debug-toolbar==3.2  # via -r requirements.in
django-filter==2.4.0      # via -r requirements.in
django-timezone-field==4.1.1  # via django-celery-beat
django==3.1.1             # via -r requirements.in, django-celery-beat, django-debug-toolbar, django-filter, django-timezone-field, djangorestframework, drf-yasg
djangorestframework==3.12.2  # via -r requirements.in, drf-yasg
drf-yasg==1.20.0          # via -r requirements.in
faker==6.4.1              # via mixer
idna==2.10                # via requests
importlib-metadata==3.4.0  # via kombu
inflection==0.5.1         # via drf-yasg
itypes==1.2.0             # via coreapi
jinja2==2.11.3            # via coreschema
kombu==4.6.11             # via celery
markupsafe==1.1.1         # via jinja2
mixer==7.1.2              # via -r requirements.in
packaging==20.9           # via drf-yasg
parameterized==0.8.1      # via -r requirements.in
pyparsing==2.4.7          # via packaging
python-crontab==2.5.1     # via django-celery-beat
python-dateutil==2.8.1    # via faker, python-crontab
pytz==2021.1              # via celery, django, django-timezone-field
redis==3.3.11             # via -r requirements.in
requests==2.25.1          # via coreapi
ruamel.yaml.clib==0.2.2   # via ruamel.yaml
ruamel.yaml==0.16.12      # via drf-yasg
six==1.15.0               # via python-dateutil
sqlparse==0.4.1           # via django, django-debug-toolbar
text-unidecode==1.3       # via faker
typing-extensions==3.7.4.3  # via importlib-metadata
uritemplate==3.0.1        # via coreapi, drf-yasg
urllib3==1.26.3           # via requests
vine==1.3.0               # via amqp, celery
zipp==3.4.0               # via importlib-metadata
psycopg2==2.8.6           # via requests

